moved access rights to own component, and new nested function
