+++ etsin_finder/frontend/.eslintrc
@@ -13,7 +13,8 @@
     "react/prefer-stateless-function": 0,
     "class-methods-use-this": 0,
     "no-param-reassign": ["error", { "props": false }],
-    "no-console": 0
+    "no-console": 0,
+    "prefer-destructuring": 0
   },
   "extends": ["airbnb"]
 }
